Refactor package to `translator` and update CLI tools
- Renamed `toolstranslator` package to `translator` and deleted legacy source files.
- Updated `translator.cli` with comprehensive commands: `install`, `doctor`, `status`, `restart`, and `clean-server`.
- Enhanced `DockerManager` with diagnostic checks, image management, and health monitoring.
- Updated Docker container name to `translator-libretranslate`.
- Refactored test imports and documentation references to use the new package name.
