--- toolstranslator/__init__.py
@@ -1,5 +1,0 @@
-"""ToolsTranslator public API."""
-
-from toolstranslator.core.translator import Translator
-
-__all__ = ["Translator"]
