--- toolstranslator/__main__.py
@@ -1,4 +1,0 @@
-from toolstranslator.cli import main
-
-if __name__ == "__main__":
-    main()
