--- toolstranslator/cli/__init__.py
@@ -1,5 +1,0 @@
-from toolstranslator.cli.app import app
-
-
-def main() -> None:
-    app()
