--- toolstranslator/docker_manager/__init__.py
@@ -1,1 +1,0 @@
-
