--- toolstranslator/file_handlers/__init__.py
@@ -1,1 +1,0 @@
-
