--- toolstranslator/file_handlers/json_handler.py
@@ -1,26 +1,0 @@
-from __future__ import annotations
-
-import json
-import os
-import tempfile
-from pathlib import Path
-from typing import Any
-
-
-class JsonHandler:
-    """JSON read/write utility with atomic write semantics."""
-
-    @staticmethod
-    def read(path: Path) -> dict[str, Any]:
-        with path.open("r", encoding="utf-8") as f:
-            return json.load(f)
-
-    @staticmethod
-    def write(path: Path, data: dict[str, Any]) -> None:
-        path.parent.mkdir(parents=True, exist_ok=True)
-        with tempfile.NamedTemporaryFile("w", encoding="utf-8", delete=False, dir=path.parent) as tmp:
-            json.dump(data, tmp, ensure_ascii=False, indent=2)
-            tmp.flush()
-            os.fsync(tmp.fileno())
-            tmp_path = Path(tmp.name)
-        tmp_path.replace(path)
